Исправления Фронта, доработка клиента

+++ Backend/ClientAPI/Services/ClientService.cs
@@ -20,7 +20,6 @@ namespace ClientAPI.Services
         enum StepsAuth
         {
             StepCheck = 0,
-            StepRegister,
             StepUpdated
         }
 
@@ -34,63 +33,77 @@ namespace ClientAPI.Services
             _tgmessage = tgmessage;
         }
 
-        private Auth_PairTokens? TokensReleased(Auth_CheckInfo check)
+        private async Task<Auth_PairTokens?> TokensReleased(Auth_CheckInfo check)
         {
             if (check.check_success == null)
                 return null;
 
-            var accessToken = _jwt.JwtTokenCreation(check.check_success);
-            var refreshToken = _jwt.RefreshTokenCreation(check.check_success);
-
-            if (_cache.CheckExistKeysStorage(check.check_success.Id, "accessTokens"))
-                _cache.DeleteKeyFromStorage(check.check_success.Id, "accessTokens");
-
-            if (_cache.CheckExistKeysStorage(check.check_success.Id, "refreshTokens"))
-                _cache.DeleteKeyFromStorage(check.check_success.Id, "refreshTokens");
-
-            _cache.WriteKeyInStorage(check.check_success.Id, "accessTokens", accessToken, DateTime.UtcNow.AddMinutes(5));
-            _cache.WriteKeyInStorage(check.check_success.Id, "refreshTokens", refreshToken, DateTime.UtcNow.AddDays(7));
-
-            _session.SetupSession(check.check_success.Id, accessToken);
-
-            Auth_PairTokens pair_tokens = new Auth_PairTokens()
-            {
-                accessToken = _cache.GetKeyFromStorage(check.check_success.Id, "accessTokens"),
-                refreshToken = _cache.GetKeyFromStorage(check.check_success.Id, "refreshTokens")
-            };
-
-            var device_out = "Неизвестного устройства";
+            var device_out = "неизвестного устройства";
 
             if (check.check_success.device == "Mobile")
-                device_out = "Мобильного устройства";
+                device_out = "мобильного устройства";
             else if (check.check_success.device == "PC")
-                device_out = "Компьютера";
+                device_out = "компьютера";
 
+            if (!_cache.CheckExistKeysStorage(check.check_success.Id, "accessTokens"))
+            {
+                var accessToken = _jwt.JwtTokenCreation(check.check_success);
+                var refreshToken = _jwt.RefreshTokenCreation(check.check_success);
 
-            _tgmessage.SendWithMarkup(check.check_success.telegram_chat_id.ToString(),
-                $"Вход с {device_out}",
-                "Это не я!", "itsnotmeQuery");
+                if (_cache.CheckExistKeysStorage(check.check_success.Id, "accessTokens"))
+                    _cache.DeleteKeyFromStorage(check.check_success.Id, "accessTokens");
 
-            _logger.LogInformation($"Пользователь {check.check_success.Id} успешно вошел!");
+                if (_cache.CheckExistKeysStorage(check.check_success.Id, "refreshTokens"))
+                    _cache.DeleteKeyFromStorage(check.check_success.Id, "refreshTokens");
 
-            return pair_tokens;
-        
+                _cache.WriteKeyInStorage(check.check_success.Id, "accessTokens", accessToken, DateTime.UtcNow.AddMinutes(5));
+                _cache.WriteKeyInStorage(check.check_success.Id, "refreshTokens", refreshToken, DateTime.UtcNow.AddDays(7));
+
+                _session.SetupSession(check.check_success.Id, accessToken);
+
+                Auth_PairTokens pair_tokens = new Auth_PairTokens()
+                {
+                    accessToken = _cache.GetKeyFromStorage(check.check_success.Id, "accessTokens"),
+                    refreshToken = _cache.GetKeyFromStorage(check.check_success.Id, "refreshTokens")
+                };
+
+                await _tgmessage.Send(check.check_success.telegram_chat_id.ToString(),
+                    $"Техническое уведомление:\nВход с {device_out}");
+
+                _logger.LogInformation($"Пользователь {check.check_success.Id} успешно вошел!");
+
+                return pair_tokens;
+            }
+            else
+            {
+                Auth_PairTokens pair_tokens = new Auth_PairTokens()
+                {
+                    accessToken = _cache.GetKeyFromStorage(check.check_success.Id, "accessTokens"),
+                    refreshToken = _cache.GetKeyFromStorage(check.check_success.Id, "refreshTokens")
+                };
+
+                await _tgmessage.Send(check.check_success.telegram_chat_id.ToString(),
+                    $"Техническое уведомление:\nЗапрос с {device_out} без обновления токенов");
+
+                return pair_tokens;
+            }
         }
 
         public async Task<Auth_PairTokens?> UserAuth(AuthAddUser dtoObj)
         {
          
-            var check = new Dictionary<StepsAuth, Auth_CheckInfo>();
+            var checkUserExist = new Dictionary<StepsAuth, Auth_CheckInfo>();
 
-            check[StepsAuth.StepCheck] = _database.CheckUser(new AuthSignIn()
+            checkUserExist[StepsAuth.StepCheck] = _database.CheckUser(new AuthSignIn()
             {
                 device = dtoObj.device, 
                 telegram_chat_id = dtoObj.chat_id
             });
 
 
-            if (check[StepsAuth.StepCheck].CheckHasSuccess())
+            if (checkUserExist[StepsAuth.StepCheck].CheckHasSuccess())
             {
+                //Обновляем профиль в сервисе если есть изменения
                 await _database.UserUpdateFromTelegram(
                     new ClientUpdate()
                     {
@@ -103,23 +116,38 @@ namespace ClientAPI.Services
                     }
                 );
 
-                check[StepsAuth.StepUpdated] = _database.CheckUser(new AuthSignIn()
+                checkUserExist[StepsAuth.StepUpdated] = _database.CheckUser(new AuthSignIn()
                 {
                     device = dtoObj.device,
                     telegram_chat_id = dtoObj.chat_id
                 });
 
-                var tokens = TokensReleased(check[StepsAuth.StepUpdated]);
+                //Выдаем токен
+                var tokens = await TokensReleased(checkUserExist[StepsAuth.StepUpdated]);
 
                 if (tokens != null)
                     return tokens;
             }
             else
             {
-                //1. Создаем заявку в кэш и ждем подтверждения
-                //2. После подтверждения добавляем в бд, удаляем из кэша заявку
-                   
-                //Если нет заявки, то создаем ее
+                throw new InvalidOperationException("user_not_found");
+            }
+
+ 
+            return null;
+        }
+
+        public async Task<string> UserRegister(AuthAddUser dtoObj)
+        {
+            var checkUserExist = _database.CheckUser(new AuthSignIn()
+            {
+                device = dtoObj.device,
+                telegram_chat_id = dtoObj.chat_id
+            });
+
+            //Если пользователя не существует в бд, создаем заявку на регу
+            if (checkUserExist.CheckHasError())
+            {
                 if (!_cache.CheckExistKeysStorage<AuthAddUser>($"register_request_{dtoObj.id}"))
                 {
                     _cache.WriteKeyInStorageObject($"register_request_{dtoObj.id}", dtoObj, DateTime.UtcNow.AddMinutes(5));
@@ -127,16 +155,19 @@ namespace ClientAPI.Services
                     await _tgmessage.SendWithMarkup(dtoObj.chat_id.ToString(), "Регистрация в сервисе \"Симбир Еда\"",
                         "Подтвердить", "registerQuery");
 
-                    throw new InvalidOperationException("register_request_created");
+                    return "register_request_created";
                 }
                 else
                 {
-                    throw new InvalidOperationException("register_request_already_exist");
+                    return "register_request_already_exist";
                 }
             }
+            else
+            {
+                throw new InvalidOperationException("user_already_exist");
+            }
 
- 
-            return null;
+           
         }
 
         public async Task<string?> ClientSignOut(string bearer_key)
